move thunks to actions and fix odd css ingredient issue
