+++ src/components/ingredient-section/ingredient-section.module.css
@@ -2,3 +2,7 @@
     display: flex;
     flex-wrap: wrap;
 }
+
+.lastOdd {
+    flex: 1 1;
+}
