+++ src/services/actions/ingredients.js
@@ -1,3 +1,6 @@
+import checkResponse from '../../utils/apiUtils';
+import BASE_URL from '../../utils/constants';
+
 export const GET_INGREDIENTS__REQUEST = 'GET_INGREDIENTS__REQUEST';
 export const GET_INGREDIENTS__SUCCESS = 'GET_INGREDIENTS__SUCCESS';
 export const GET_INGREDIENTS__FAILURE = 'GET_INGREDIENTS__FAILURE';
@@ -16,3 +19,45 @@ export const CURRENT_ITEM_CLOSE = 'CURRENT_ITEM_CLOSE';
 
 export const SET_TOTAL_PRICE = 'INCREASE_TOTAL_PRICE';
 export const RESET_TOTAL_PRICE = 'DECREASE_TOTAL_PRICE';
+
+export function getIngredients() {
+  return function (dispatch) {
+    dispatch({ type: GET_INGREDIENTS__REQUEST });
+    return fetch(`${BASE_URL}/ingredients`)
+      .then((res) => checkResponse(res))
+      .then((data) => {
+        dispatch({ type: GET_INGREDIENTS__SUCCESS, payload: data });
+      })
+      .catch((err) => {
+        console.log(err);
+        dispatch({ type: GET_INGREDIENTS__FAILURE, payload: err });
+      });
+  };
+}
+
+export function createOrderRequest(constructorIngredients) {
+  return function (dispatch) {
+    // Start the API call by dispatching a request action
+    dispatch({ type: POST_ORDER__REQUEST });
+
+    return fetch(`${BASE_URL}/orders`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+        authorization: localStorage.getItem('token'),
+      },
+      body: JSON.stringify({
+        ingredients: constructorIngredients,
+      }),
+    })
+      .then((res) => checkResponse(res))
+      .then((res) => {
+        dispatch({ type: POST_ORDER__SUCCESS, payload: res });
+        return res;
+      })
+      .catch((err) => {
+        dispatch({ type: POST_ORDER__FAILURE, payload: err });
+        console.log(err);
+      });
+  };
+}
