+++ src/utils/apiUtils.js
@@ -1,0 +1,4 @@
+const checkResponse = (res) => (res.ok ? res.json() : res.json()
+  .then((err) => Promise.reject(err)));
+
+export default checkResponse;
