+++ src/utils/constants.js
@@ -1,0 +1,3 @@
+const BASE_URL = 'https://norma.nomoreparties.space/api';
+
+export default BASE_URL;
